Deprecate scripts/pyse.py in favor of sourcefinder/utility.cli.py
